+++ src/hotaru/__main__.py
@@ -1,5 +1,18 @@
-import sys
+def app():
+    import os
 
-from .console.app import app
+    os.environ["TF_CPP_MIN_LOG_LEVEL"] = "2"
 
-sys.exit(app())
+    import tensorflow as tf
+
+    tf.get_logger().setLevel("ERROR")
+
+    from .console.main import main
+
+    import sys
+    sys.exit(main())
+    return main()
+
+
+if __name__ == "__main__":
+    app()
