+++ src/hotaru/console/run/clean.py
@@ -14,8 +14,8 @@ from .options import radius_options
     click.Option(["--stage", "-s"], type=int),
     click.Option(["--footprint-tag"]),
     click.Option(["--footprint-stage"], type=int),
-    click.Option(["--gauss"], type=float),
     *radius_options(),
+    click.Option(["--gauss"], type=float),
     click.Option(["--thr-area-abs"], type=click.FloatRange(0.0)),
     click.Option(["--thr-area-rel"], type=click.FloatRange(0.0)),
     click.Option(["--thr-sim"], type=click.FloatRange(0.0, 1.0)),
