+++ src/hotaru/console/run/temporal.py
@@ -4,8 +4,8 @@ import tensorflow as tf
 from ...evaluate.summary import write_spike_summary
 from ...train.temporal import TemporalModel
 from ..base import run_command
-from ..progress import ProgressCallback
 from .options import model_options
+from .progress import ProgressCallback
 
 
 @run_command(
